activity/item navigation completed
