+++ package.json
@@ -12,8 +12,8 @@
     "bootstrap": "^4.6.0",
     "connected-react-router": "^6.8.0",
     "dayspan": "^1.1.0",
-    "formik": "^2.2.6",
     "expr-eval": "^2.0.2",
+    "formik": "^2.2.6",
     "i18next": "^19.9.2",
     "i18next-browser-languagedetector": "^6.0.1",
     "i18next-xhr-backend": "^3.2.2",
@@ -31,6 +31,7 @@
     "react-fontawesome": "^1.7.1",
     "react-hook-form": "^6.9.2",
     "react-i18next": "^11.8.10",
+    "react-markdown": "^6.0.2",
     "react-redux": "^7.2.1",
     "react-redux-form": "^1.16.14",
     "react-router-dom": "^5.2.0",
@@ -41,6 +42,7 @@
     "redux-logger": "^3.0.6",
     "redux-persist": "^6.0.0",
     "redux-thunk": "^2.3.0",
+    "remark-gfm": "^1.0.0",
     "reselect": "^4.0.0",
     "styled-components": "^5.2.0"
   },
