+++ src/App.js
@@ -1,36 +1,35 @@
-import React from 'react'
-import { Container } from 'react-bootstrap'
-import { ConnectedRouter } from 'connected-react-router'
-import { Switch, Route } from 'react-router-dom'
-import { useSelector } from 'react-redux'
+import React from 'react';
+import { Container } from 'react-bootstrap';
+import { ConnectedRouter } from 'connected-react-router';
+import { Switch, Route } from 'react-router-dom';
+import { useSelector } from 'react-redux';
 
-import { history } from './store'
-import { userInfoSelector } from './state/user/user.selectors'
+import { history } from './store';
+import { userInfoSelector } from './state/user/user.selectors';
 
-import Footer from './components/Base/Footer'
-import NavBar from './components/Base/Navbar'
-import Login from './components/Login'
-import SignUp from './components/Signup'
-import ForgotPassword from './components/ForgotPassword'
-import Profile from './components/Profile'
-import ChangePassword from './components/ChangePassword'
-import Landing from './components/Landing'
-import AppletList from './components/AppletList'
-import Invitation from './components/Invitation/Invitation'
-import AcceptInvitation from './components/Invitation/AcceptInvitation'
-import DeclineInvitation from './components/Invitation/DeclineInvitation'
-import AppletParentRoute from './components/AppletParentRoute'
-import AppletDashboard from './components/AppletParentRoute/AppletDashboard'
-import SetPassword from './components/Setpassword'
-import { Consent } from './components/Consent/index'
-import Screens from './components/Screens'
+import Footer from './components/Base/Footer';
+import NavBar from './components/Base/Navbar';
+import Login from './components/Login';
+import SignUp from './components/Signup';
+import ForgotPassword from './components/ForgotPassword';
+import Profile from './components/Profile';
+import ChangePassword from './components/ChangePassword';
+import Landing from './components/Landing';
+import AppletList from './components/AppletList';
+import Invitation from './components/Invitation/Invitation';
+import AcceptInvitation from './components/Invitation/AcceptInvitation';
+import DeclineInvitation from './components/Invitation/DeclineInvitation';
+import AppletDashboard from './components/AppletParentRoute/AppletDashboard';
+import SetPassword from './components/Setpassword';
+import { Consent } from './components/Consent/index';
+import Screens from './components/Screens';
 
-import './App.css'
+import './App.css';
 
 /**
  * Main Component of the Application - Routes to specific components based on the Path.
  */
-function App() {
+const App = () => {
   const user = useSelector(userInfoSelector)
 
   return (
@@ -90,4 +89,4 @@ function App() {
   )
 }
 
-export default App
+export default App;
