+++ src/components/Item/index.js
@@ -3,21 +3,21 @@ import { Formik, Form } from 'formik';
 
 // Widgets
 import Radio from '../../widgets/Radio';
-import TextInput from '../../widgets/TextInput';
 import Checkbox from '../../widgets/Checkbox';
+import TextInput from '../../widgets/TextInput';
 import Slider from '../../widgets/Slider/index';
 
 import "./style.css";
 
 const Item = (props) => {
   const { data, type, handleSubmit } = props;
-  
+
   const widget = (handleChange) => {
     switch (type) {
-      case "radio":
-        return <Radio {...props} handleChange={handleChange} />;
       case "checkbox":
         return <Checkbox {...props} handleChange={handleChange} />;
+      case "radio":
+        return <Radio {...props} handleChange={handleChange} />;
       case "textinput":
         return <TextInput {...props} handleChange={handleChange} />;
       case "slider":
