+++ src/index.css
@@ -60,3 +60,19 @@ code {
 .question {
   text-align: center;
 }
+
+.question img {
+  width: 250px;
+  max-height: 200px;
+  margin-bottom: 5px;
+  float: left;
+  border-radius: 0.25rem;
+  border: 1px solid #dee2e6;
+}
+
+.question video {
+  width: 250px;
+  float: left;
+  border-radius: 0.25rem;
+  border: 1px solid #dee2e6;
+}
