+++ src/state/app/app.constants.js
@@ -11,7 +11,6 @@ export default {
   TOGGLE_MOBILE_DATA_ALLOWED: 'TOGGLE_MOBILE_DATA_ALLOWED',
   REDIRECT_URL: 'REDIRECT_URL',
   GET_INVITATION: 'GET_INVITATION',
-  GET_APPLETS: 'GET_APPLETS',
   ACCEPT_INVITATION: 'ACCEPT_INVITATION',
   DECLINE_INVITATION: 'DECLINE_INVITATION',
 }
