+++ src/state/app/app.reducer.js
@@ -10,8 +10,6 @@ export const initialState = {
    * @type {string}
    */
   apiHost: config.defaultApiHost,
-  applets: [],
-  responses: [],
 
   /**
    * The current skin (theme) for the app.
@@ -97,18 +95,6 @@ const AppSlice = createSlice({
     toggleMobileDataAllowed: (state, action) => { state.mobileDataAllowed = !state.mobileDataAllowed },
     setRedirectUrl: (state, action) => { state.redirectUrl = action.payload },
   },
-  extraReducers: {
-    [`${APP_CONSTANTS.GET_APPLETS}/pending`]: (state, action) => { state.loading = true; state.error = null },
-    [`${APP_CONSTANTS.GET_APPLETS}/fulfilled`]: (state, action) => {
-      state.loading = false;
-      state.error = null;
-      state.applets = action.payload;
-    },
-    [`${APP_CONSTANTS.GET_APPLETS}/rejected`]: (state, action) => {
-      state.loading = false;
-      state.error = action.error.message;
-    },
-  }
 })
 
 export default AppSlice.reducer;
