+++ src/state/applet/applet.reducer.js
@@ -9,6 +9,8 @@ export const initialState = {
    * @type {object}.
    */
   applets: [],
+  applet: {},
+  responses: [],
   isReminderSet: false,
   scheduleUpdated: false,
   currentTime: new Date(),
@@ -30,10 +32,25 @@ const appletSlice = createSlice({
   initialState,
   reducers: {
     clearApplets: () => initialState,
+    selectApplet: (state, action) => { state.applet = action.payload },
+    selectActivity: (state, action) => { state.activityAccess = action.payload },
 
   },
+  extraReducers: {
+    [`${APPLET_CONSTANTS.GET_APPLETS}/pending`]: (state, action) => { state.loading = true; state.error = null },
+    [`${APPLET_CONSTANTS.GET_APPLETS}/fulfilled`]: (state, action) => {
+      state.loading = false;
+      state.error = null;
+      state.applets = action.payload;
+    },
+    [`${APPLET_CONSTANTS.GET_APPLETS}/rejected`]: (state, action) => {
+      state.loading = false;
+      state.error = action.error.message;
+    },
+  }
+
 })
 
 export default appletSlice.reducer;
 
-export const { clearUser } = appletSlice.actions;
+export const { clearApplets, selectApplet, selectActivity } = appletSlice.actions;
